requirements.txt
streamlit
nltk
scikit-learn
joblib
numpy
pandas
faiss-cpu
sentence-transformers
groq
newspaper3k